+++ ironflow/ironflow/gui.py
@@ -17,9 +17,9 @@ __date__ = "May 10, 2022"
 
 import ipywidgets as widgets
 from IPython.display import display, HTML
-from ryven.ironflow.model import HasSession
-from ryven.ironflow.node_interface import NodeController, NodePresenter
-from ryven.ironflow.canvas_widgets import FlowCanvas
+from ironflow.ironflow.model import HasSession
+from ironflow.ironflow.node_interface import NodeController, NodePresenter
+from ironflow.ironflow.canvas_widgets import FlowCanvas
 
 from typing import Optional
 from ryvencore import Session
