+++ ironflow/ironflow/node_interface/control.py
@@ -4,7 +4,7 @@
 
 from __future__ import annotations
 
-from ryven.ironflow.node_interface.base import NodeInterfaceBase
+from ironflow.ironflow.node_interface.base import NodeInterfaceBase
 from IPython.display import display
 import ipywidgets as widgets
 import numpy as np
@@ -14,8 +14,8 @@ import base64
 
 from typing import TYPE_CHECKING, Callable, Optional
 if TYPE_CHECKING:
-    from ryven.ironflow.gui import GUI
-    from ryven.NENV import Node
+    from ironflow.ironflow.gui import GUI
+    from ironflow.NENV import Node
 
 __author__ = "Joerg Neugebauer, Liam Huber"
 __copyright__ = (
