+++ ironflow/nodes/pyiron/nodes.py
@@ -2,7 +2,7 @@
 # Copyright (c) Max-Planck-Institut für Eisenforschung GmbH - Computational Materials Design (CM) Department
 # Distributed under the terms of "New BSD License", see the LICENSE file.
 
-from ryven.NENV import export_nodes
+from ironflow.NENV import export_nodes
 
 import sys
 import os
