--- ryven/ironflow/__init__.py
@@ -1,2 +1,0 @@
-from ryven.ironflow.gui import GUI
-from ryven.NENV import Node, NodeInputBP, NodeOutputBP, dtypes
